# docs/swagger.yaml
basePath: /
definitions:
  dto.AddReview:
    properties:
      dt_orderid:
        type: integer
      rating:
        type: integer
    required:
      - dt_orderid
      - rating
    type: object
  dto.CreateMenuOrder:
    properties:
      menu_id:
        type: integer
      product_size_id:
        type: integer
      product_type_id:
        type: integer
      qty:
        type: integer
    type: object
  dto.CreateOrder:
    properties:
      menus:
        description: Status   string            `json:"status" binding:"required"`
        items:
          $ref: "#/definitions/dto.CreateMenuOrder"
        type: array
      payment_id:
        type: integer
      shipping:
        type: string
    required:
      - menus
      - payment_id
      - shipping
    type: object
  dto.DetailItemResponse:
    properties:
      image:
        type: string
      item_name:
        type: string
      product_size:
        type: string
      product_type:
        type: string
      qty:
        type: integer
      subtotal:
        type: string
    type: object
  dto.DetailOrderResponse:
    properties:
      address:
        type: string
      date_order:
        type: string
      detail_item:
        items:
          $ref: "#/definitions/dto.DetailItemResponse"
        type: array
      fullname:
        type: string
      order_id:
        type: string
      payment_method:
        type: string
      phone:
        type: string
      shipping:
        type: string
      status:
        type: string
      total:
        type: string
    type: object
...
  dto.History:
    properties:
      date:
        type: string
      order_id:
        type: string
      status:
        type: string
      total:
        type: number
    type: object
...
  dto.ProductType:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
...
  dto.ResponseError:
    properties:
      errors:
        example: failed get data
        type: string
      message:
        example: Failed get data
        type: string
      status:
        example: Error
        type: string
    type: object
  dto.ResponseSuccess:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
...
  dto.UpdateStatusOrder:
    properties:
      order_id:
        type: string
      status:
        type: string
    required:
      - order_id
      - status
    type: object
...
host: 192.168.50.221:8080
info:
  contact: {}
  description: Solid Coffee Backend RESTful API
  title: Solid Coffee Backend
  version: "1.0"
paths:
  /admin/orders/:
    get:
      parameters:
        - description: Page Start
          in: query
          name: page
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/dto.ProductType"
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Get all order
      tags:
        - Admin Order Management
    patch:
      consumes:
        - application/json
      parameters:
        - description: Update status order
          in: body
          name: orders
          required: true
          schema:
            $ref: "#/definitions/dto.UpdateStatusOrder"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Update status
      tags:
        - Orders
...
  /orders/:
    post:
      consumes:
        - application/json
      parameters:
        - description: Create order
          in: body
          name: movies
          required: true
          schema:
            $ref: "#/definitions/dto.CreateOrder"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Create order
      tags:
        - Orders
  /orders/history:
    get:
      parameters:
        - description: Page Start
          in: query
          name: page
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/dto.History"
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Get all history
      tags:
        - orders
  /orders/history/{id}:
    get:
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.DetailOrderResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Get detail history by id
      tags:
        - orders
  /orders/review/:
    post:
      consumes:
        - application/json
      parameters:
        - description: Add review
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.AddReview"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/dto.ResponseSuccess"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/dto.ResponseError"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/dto.ResponseError"
      security:
        - BearerAuth: []
      summary: Add review to order
      tags:
        - Admin Order Management
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"